Issue #61 : Fix unload and close broker error (#77)
Fix issue #61
We did not handle bundle unload very well. So in the situation that a topic is unload, the original kafka connection is still active, and topics still cached in KoP, this will cause error.
In this change, we track the topic loading status by register a mock producer in to PersistentTopic. when topic is unload, it will trigger the connection close.
Add new unit tests, make sure all tests passed.