Cache Pulsar topic lookup result to improve performance (#80)
current code will cause too much lookup. This PR try to cache the topic lookup result for topics that served by this broker.  For the topics that not served by this broker, we should not cache it, since it is not able to track the topic reloading, and will cause conflict.