Issue 113: fix request order not align with response order issue (#114)
Fix #113 
In current request handler. we passed in the request and after some handles return a responseFuture.
If several request comes in at same time, the handling time difference will cause the return future order different with the order that command comes in. And this will cause request handling order different.

This error is seen in the KafkaMessageOrderTest:
- client send 10 messges, while first 9 messages batched together, and the last one not batched;
- the above 2 request comes into handleProduceRequest, but the second request returned earlier than the first.

The fix is:
- add request response future in order at the time that request comes in, and call related handlers later.
- add produce request in order, and handle one by one in each channel. 

* add request response future at the time request comes
* make produce request in order
* change cursor delete to async in KafkaTopicConsumerManager
* add some logs of offset