Auto set listeners when it is not set in Protocol Handler. (#121)
The default value for `listeners` is currently `PLAINTEXT://127.0.0.1:9092`, in which, the ip is the advertise address of Pulsar.
In K8S env. user usually need set `listeners` as config map of yaml file, and passed into pods,  it is not easy to get pod's "advertise address" and put it into yaml file.
This change will config the `listeners` to value `PLAINTEXT://{advertise.address}:9092` if it is not set. So most of the case user not need to config this parameter. 
This change only change the code path of protocol handler. Not change the KafkaBroker and KafkaStandalone, since they will be removed.

* auto config default listener for protocolHandler

* add comments in .conf