remove redundance reference to get rid of native memory leak when getting a lot of requests. (#232)
KafkaHeaderAndRequest contains native memory references. When getting a lot of requests from kafka client, for example 10 million per seconds, those redundance references will stop gc collection and memory get increased.
It would be tricky to provide unit test for this patch and this pr seems to be a minor change, so I didn't write test case for it.