Set controller to broker itself when metadata request is processed (#242)
KoP needs no controller id because it's used for metadata operations in Kafka. So it simply sets controller id field to -1 when metadata request is processed.

However, before a Kafka's AdminClient send a metadata related request like CreateTopics, it sent a metadata request and try to get a valid controller broker from the controller id field of metadata response. If the controller id is invalid, the AdminClient will send metadata request infinitely.

Since each broker can manage metadata, this PR change metadata response's controller id to the self node's controller id, which is generated by a hash function.

In addition, this PR adds a test to verify Kafka's AdminClient won't stuck at the CreateTopics method which is not supported currently.