Fill the offset field of PartitionResponse (#245)
Fixes #230

The offset field is not filled when a PRODUCE request is processed yet. It causes that the offset what Kafka producer get from send()'s callback is always -1, which may cause some problems like #241 mentioned.

This PR fills the offset field and add a unit test to verify if the right offset is set.