Support CreateTopics request (#244)
Master issue: #241

This PR add support for Kafka CreateTopics request so that topics can be created by Kafka's AdminClient. Related tests are added to verify

The validateOnly field introduced from CreateTopics request v1 is not processed. The validateOnly field only checks the topics can be created as specified but not creates anything. It's not a important feature currently so leave it as a future task.


* Add handler for CreateTopics request

* Fix wrong numTopics and default namespace

* Add test for CreateTopics request handler

* Handle the corner case when topic name is invalid

* Add this to make code more readable