Remove namespace prefix for topics in the default namespace (#254)
Master issue: #241

Currently if a Kafka Client tries to list all topics, the topic name in response will contain the namespace prefix like "persistent://public/default". However, for topics in the default namespace, they shouldn't contain the prefix because Kafka Client would like to use the short topic name. So this PR will remove the prefix of topics in the default namespace.


* Add a method to remove the default namespace prefix

* Use short topic name if the topic is in the default namespace

* Change variable name to avoid conflict

* Fix simpleProduceAndConsumeWithPulsarAuthed test error