Fix message disorder (#253)
Fixes #247

The original implementation of message publishing cannot guarantee that a topic partition's pending writes are completed in order. This PR is to fix the issue by refactoring the handler for Produce requests.

The basic steps of handling a single partition's produce request are:

Get PersistentTopic from topic manager;
Convert Kafka's MemoryRecords to Pulsar's ByteBuf;
Call PersistentTopic#publishMessages to write ByteBuf to BK asynchronously.
This PR adds a PendingProduce class to compose pending step 1 and 2 to a CompletableFuture, then add the PendingProduce object to a queue (PendingProduceQueue) which is associated with the partition name in a map named pendingProduceQueueMap. When the step 1 and 2 are completed, the queue will try to remove all ready PendingProduce objects in the head and call PersistentTopic#publishMessages. We use synchronized keyword to make the remove operation thread safe.

Therefore, the first two steps can be executed in parallel, but the third step is executed in order.

In addition, this PR fixes the current message order test by using different batch size. In the original test, all messages are batched to a single batch, so the disorder never happens because there's only one batch.


* Refactor handleProduceRequest to fix message disorder issue

* Fix existed message order test

* Synchronize PendingProduce#publishMessages

* Test message order with different batch.size config