Add the test framework for Kafka Streams (#272)
This PR is the first step to port Kafka Streams' [integration tests](https://github.com/apache/kafka/tree/2.0/streams/src/test/java/org/apache/kafka/streams/integration) to KoP.

There're two main differences from the original tests in Kafka:
1. Kafka's tests use a lot of util methods from different modules, so this PR adds a `TestUtils.java` to collect all necessary util methods.
2. This PR abstracts a base class for reused variables and methods.

Based on the framework, this PR ports the tests for `GlobalKTable`. It should be noted that in `shouldRestoreGlobalInMemoryKTableOnRestart` we sleep for 1 second after `startStreams()` because it may take a few milliseconds to wait streams started in KoP.

Finally, the old test for word count example is deleted since it will be redundant after tests are ported.