Fix topic auto created when using Kafka AdminClient to get metadata (#274)
Fixes #273

When KoP handles metadata requests, it only checks the config allowAutoTopicCreation but not the same field in the metadata request. Then it causes that AdminClient#describeTopics returns a success result even if the topics don't exist and topics will be created automatically.

So this PR adds the check for metadata request's allowAutoTopicCreation field and the test for AdminClient#describeTopics.