Fix flaky testCreateInvalidTopics (#361)
The `testCreateInvalidTopics` sometimes failed in CI environment, like

```
Error:  testCreateInvalidTopics(io.streamnative.pulsar.handlers.kop.KafkaRequestHandlerTest)  Time elapsed: 5.022 s  <<< FAILURE!
java.lang.AssertionError: expected [true] but found [false]
	at org.testng.Assert.fail(Assert.java:96)
	at org.testng.Assert.failNotEquals(Assert.java:776)
	at org.testng.Assert.assertTrue(Assert.java:44)
	at org.testng.Assert.assertTrue(Assert.java:54)
	at io.streamnative.pulsar.handlers.kop.KafkaRequestHandlerTest.testCreateInvalidTopics(KafkaRequestHandlerTest.java:359)
```

However here we only use `assertTrue` to check the cause type of the `ExecutionException` but not printed it. So this PR adds error logs for debugging.

Normally it should returns a `TimeoutException`. However, the `KafkaAdmin` could also generate an `UnknownServerException` sometimes, whose error message is just what the server's response is. Here's a sample output of a CI failure:

> Failed to create topics: {xxx/testCreateInvalidTopics-0=1} caused by org.apache.kafka.common.errors.UnknownServerException: Invalid short topic name 'xxx/testCreateInvalidTopics-0', it should be in the format of <tenant>/<namespace>/<topic> or <topic>

The exception message is from `KopTopic#expandToFullName` that is thrown in `AdminManager#createTopicsAsync`.

So this PR looses the check of exception type to make the test pass.