Support OAUTHBEARER mechanism for KoP (#400)
This PR introduces OAUTHBEARER mechanism to KoP.

Two config items are added. One is the server callback handler to validate OAUTHBEARER token, the other is the properties configuration file that contains the server callback handler's configs. The default server callback handler is the same as Kafka, which validates unsecured JSON Web Tokens.

In addition, the authenticator is refactored that now it uses different `SaslServer` for different mechanisms to perform authentication.

Related tests are added to verify these configs and the default server callback handler:
- `KafkaServerConfiguration#testGetKopOauth2Configs`: unit test for `getKopOauth2Properties` method.
- `SaslOAuthBearerTest`: test `OAUTHBEARER` SASL mechanism with Kafka's default server callback handler.
- `CustomOAuthBearerCallbackHandlerTest`: test custom `AuthenticateCallbackHandler`.