[Transaction] Add transaction log key and transaction log value (#424)
# Motivation

Transaction log key and value are the preconditions of the transaction log.

The transaction log key will be encoded to bytes and set in message key bytes.
The transaction log value will be encoded to bytes and set in message value bytes.

The transaction log key is used to store the transactional id, the transaction log value is used to store the transaction metadata, such as state, topic partitions, start time, etc. If the broker stopped or crashed, we could recover the transaction metadata from the transaction log.