Fix NPE in MessageFetchContext (#480)
Fixes #479 

The NPE is introduced from https://github.com/streamnative/kop/pull/467. When the message read stats is recorded, the `recycle()` method of `MessageFetchContext` may have already been called and the `requestHandler` will become null after that, then the NPE happens.

This PR uses a simple way to fix it by extending the lifetime of `messageReadStats`. We can verify the fix by checking the output of a test like `BasicEndToEndKafkaTest#testNullValueMessages` and no NPE error logs will be found.