Refactor FETCH request handler to reduce GC (#538)
### Motivation

When a consumer sends FETCH requests frequently, for example, polling in a tight loop,

```java
while (true) {
    ConsumerRecords<String, byte[]> records = consumer.poll(Duration.ofMillis(100));
    // process records quickly
}
```
the heap memory increasing quickly and GC will happen frequently about once per minute. The frequency is not related to the throughput but the count of FETCH requests.

It's because each time a FETCH request arrived, KoP will create a `MessageFetchContext`. Though it's allocated in a Netty recycler, there're a lot of small objects that are still allocated from heap memory. There're a lot of wait all operations, for example, after all `PersistentTopic`s of each partition were get, it collected all those `PersistentTopic`s to a list and retrieve partition's associated data from a `Map`.

We should process each partition in parallel and avoid collecting the middle result to a list and map.

### Modifications

Refactor the `MessageFetchContext`, only collect the results to a list after the final `PartitionData` of each partition was done.

Besides, this PR removes some implementation details.
1. Checks for `minBytes` and `maxBytes`. Because the existed check is wrong. The default `minBytes` is 1 so `allSize > minBytes || allSize > maxBytes` is always true.  The `maxBytes` check is meaningless.
2. This PR only tries to read `maxReadEntriesNum` entries for only once. It's the same reason as previous.
3. Remove the `TOTAL_MESSAGE_READ` metrics because we don't wait all partitions' messages read anymore.
4. Remove the handle of FETCH timeout process, because it was still wrong before that it didn't count for `asyncReadEntries`, which is more likely to be blocked.

For the read bytes limit like `max.partition.fetch.bytes` and `fetch.max.bytes`, we should open a new PR to implement it correctly.