Make DecodeResult list thread safe (#547)
Fixes #546 

### Motivation

After the FETCH handler refactor in #538 , there's a thread safety problem when accessing `decodeResults` because each `DecodeResult` object is created in the callback of each partition's future and these callbacks may be not executed in the same thread. So when there're many partitions, accessing `decodeResults` is not thread-safe.

Before #538, the implementation just waits until all partitions' read completed, then collects the `DecodeResult` to a list. So using an `ArrayList` was thread safe before.

### Modifications

Use `ConcurrentLinkedQueue` as the container type of `decodeResults` for thread safety.