Handle group name was failed to be written to ZK gracefully (#554)
### Motivation

Currently KoP writes group name and the associated client's host and id to ZK so that consumer stats could be updated while FETCH request contains no group info. However, there is something wrong with ZooKeeperUtils.

First, when the `groupIdStoredPath` already existed, the write to ZK may fail with `NodeExists` because of race condition.

Second, if the data was failed to be written to ZK, every time a FETCH request comes, a ZK read operation will be performed. It's because `ZooKeeperUtils#getData` returns null if the node doesn't exist and a null value cannot be put into a `Map`. In this case, `ConcurrentHashMap#computeIfAbsent` will always execute the second argument to create a value.

### Modifications

- Create groupIdZooKeeperPath only when KoP starts, and only log the error message if ZK write failed.
- When `ZooKeeperUtils#getData` failed, returns an empty string instead of null so that if the group name was failed to be written to ZK, the empty returned value would be cached.