[Docs] Add more details for SSL keys (#569)
Currently the docs related to SSL keys is not friendly to users that are not familiar with `keytool` and `openssl` commands. For some commands we need to specify the password, while for some other commands we need to input the password we have specified before.

In addition, this PR removes the redundant blanks of the line tail.