Support Kafka consumer for Pulsar format messages when entryFormat is kafka (#632)
### Motivation

There's a limit when `entryFormat=kafka` is configured that Kafka client can only interact with Kafka client but not Pulsar client. This PR intends to support Kafka consumer interacts with Kafka producer.

BTW, to support Kafka producer interacts with Kafka consumer, we need a feature from Pulsar's side.

### Modifications

Add a method `decodePulsarEntryToKafkaRecords` to convert Pulsar's metadata and payload to Kafka's `MemoryRecords`. Since it's basically migrated from the current `PulsarEntryFormatter`'s implementation, both `KafkaEntryFormatter` and `PulsarEntryFormatter` will reuse this method.

In `KafkaEntryFormatter#decode`, for each entry, first parse the `entry.format` property that is set in `KafkaEntryFormatter#encode`, if it's not `kafka`, treat it as a message of Pulsar format and use `decodePulsarEntryToKafkaRecords` to decode it.

In addition, this PR also fixes some bugs of the original Pulsar format decode implementation:
1. Use `isPartitionKeyB64Encoded()` instead of `hasPartitionKeyB64Encoded()` to check if the key is Base64 encoded. Because currently Pulsar always set this field false so that `hasPartitionKeyB64Encoded()` returns true but the key is not Base64 encoded.
2. Don't return the `ByteBuffer` passed to `MemoryRecordsBuilder`'s constructor as the returned value. Because if `MemoryRecordsBuilder`'s internal `ByteBuffer` doesn't have enough space to contain more bytes, it will increased until the max bytes limit, then the outer `ByteBuffer` won't be the actual `ByteBuffer`. This bug was not exposed yet just because the initial buffer size is 1MB, which is great enough usually.

Finally, this PR adds two tests to verify it. One is to verify Pulsar producer and Kafka consumer for messages with or without keys or headers. The other is to verify mixed (Pulsar & Kafka) producers and Kafka consumer for messages with keys and headers.