Avoid creating a lot of MetadataCache instances (#663)
### Motivation

When there're a lot of topics, there could be a lot of `MetadataCache` instances in memory.

![image](https://user-images.githubusercontent.com/18204803/129903767-4b2f3961-c0e1-4ad5-88b0-7451ed41393a.png)

This bug was introduced from https://github.com/streamnative/kop/pull/488 because the Pulsar side adopted a new class `MetadataCache` as ZK cache.  However, each time `MetadataStore#getMetadataCache` is called, a new `MetadataCache` instance will be created. In KoP, it means that each time a topic lookup is performed, a new `MetadataCache` instance will be created.

### Modifications

Only call `MetadataStore#getMetadataCache` once when KoP starts and reuse the same `MetadataCache` instance for each `KafkaRequestHandler` instance.