[FEATURE] Support non-partitioned Topic (#690)
Fix #674

## Motivation
Some users want to use exists non-partitioned topic in KoP, so we need support the feature.

## Modifications
**Note**: Current don't support transaction.

When KoP handle `METADATA` request, if the topic is non-partitioned topic, then treat as a single partitioned topic. And when we `KafkaTopicManager#getTopic`, add a fallback to get non-partitioned topic. Also when namespace `load` or `unload`, we need clear the cache.