Update docs for multiple listeners and non-partitioned topic support (#760)
### Motivation

https://github.com/streamnative/kop/pull/690 supports non-partitioned topics and https://github.com/streamnative/kop/pull/742 supports multiple listeners and marks some configs as deprecated, so the current documents are outdated. 

### Modification
Even if KoP supports non-partitioned topics now, it's still better to configure `allowAutoTopicCreationType` with `partitioned` by default. This PR modified the explanation for the reason.

This PR also adds docs for how to enable multiple listeners and updates the related configuration docs.

In addition, https://github.com/streamnative/kop/pull/742 marks `kafkaListenerName` as deprecated, but it also makes this config not work anymore. So this PR removes the config and related tests.