Remove and close TCM according to client address (#794)
Fixes #773 

### Motivation

See https://github.com/streamnative/kop/issues/773 for details.

### Modifications

Add a `removeAndCloseByAddress` method to be called instead of `close` in `KafkaTopicManager#close`. Rename the current `removeAndClose` to `removeAndCloseByTopic` to differ these two methods. As for the `cursorExpireTask`, only cancel it in `KafkaProtocolHandler#close`.

Add a test to verify the close of any producer doesn't affect the TCM cache, and the close of any consumer only affects the cache associated with the consumer's address.