Fix testProduceConsumeMultiLedger might fail when consumer receive message timeout (#843)
Fixes: #842 
## Motivation
When `consumer.receive(100, TimeUnit.MILLISECONDS)` timeout, the message is null, this is a normal situation, because sometimes there are network fluctuations or other situation. We should skip the null messages.

## Modification
Skip all null messages, since we added timeout in test method, it won't cause the test get stuck be here.